// src/frontend/angular/src/app/core/api/generated/models/Web_Dtos_UpdateUserRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Web_Dtos_Gender } from './Web_Dtos_Gender';
import type { Web_Dtos_UserRole } from './Web_Dtos_UserRole';
export type Web_Dtos_UpdateUserRequest = {
    name?: string | null;
    phone?: string | null;
    gender?: Web_Dtos_Gender;
    role?: Web_Dtos_UserRole;
    password?: string | null;
};